Feat: new components
It were created 6 components which these are; AccordionInfoProduct.jsx, AccordionQA.jsx, AccordionRateProduct.jsx, CardProduct.jsx, CardProductAdaptive.jsx with its scss file and ImageGallery.jsx. These components were implemented in the component Product.jsx . Also were some changes in the Product.scss and were added new data in the file static.data.js
