+++ tony-shop/src/atoms/product-description/Product.jsx
@@ -1,538 +1,71 @@
 import React, { useState } from 'react';
 import { useParams } from 'react-router-dom';
 // component
+import ImageGallery from '@molecules/image-Gallery/ImageGallery';
+import CardProduct from '@molecules/card/cardProduct-details/CardProduct';
+import CardProductAdaptive from '@molecules/card/cardProduct-details/CardProduct-adaptive/CardProductAdaptive';
+import AccordionInfoProduct from '@molecules/accordion/accordion-info-product/AccordionInfoProduct';
+import AccordionQA from '@molecules/accordion/accordion-q&a/AccordionQA';
+import AccordionRateProduct from '@molecules/accordion/accordion-rate-product/AccordionRateProduct';
 import SliderGrid from '@molecules/slider/sliderGrid/SliderGrid';
-
+// MUI component
 import Container from '@mui/material/Container';
-
-// cosas para el menu
-import Typography from '@mui/material/Typography';
-import Box from '@mui/joy/Box';
-import Card from '@mui/joy/Card';
-import CardContent from '@mui/joy/CardContent';
-import CardActions from '@mui/joy/CardActions';
-import Rating from '@mui/material/Rating';
-import IconButton from '@mui/material/IconButton';
-import AddCircleOutlineIcon from '@mui/icons-material/AddCircleOutline';
-import RemoveCircleOutlineIcon from '@mui/icons-material/RemoveCircleOutline';
-import Input from '@mui/joy/Input';
-import Button from '@mui/material/Button';
-import ShoppingCartIcon from '@mui/icons-material/ShoppingCart';
-import Favorite from '@mui/icons-material/Favorite';
 import Stack from '@mui/material/Stack';
-
-// cosas para la img
-import AspectRatio from '@mui/joy/AspectRatio';
-import Avatar from '@mui/joy/Avatar';
-
-// cosas para la info
-import Accordion from '@mui/material/Accordion';
-import AccordionSummary from '@mui/material/AccordionSummary';
-import AccordionDetails from '@mui/material/AccordionDetails';
-import ExpandMoreIcon from '@mui/icons-material/ExpandMore';
-import Divider from '@mui/joy/Divider';
-
-// cosas para la tabla
-import List from '@mui/material/List';
-import ListItem from '@mui/material/ListItem';
-import ListItemText from '@mui/material/ListItemText';
-import ListItemButton from '@mui/material/ListItemButton';
-import ListItemIcon from '@mui/material/ListItemIcon';
-
-// cosass para las preguntas
-import TextField from '@mui/material/TextField';
-import SubdirectoryArrowRightIcon from '@mui/icons-material/SubdirectoryArrowRight';
-
+import Typography from '@mui/material/Typography';
 // static file
-import { rangeCalification } from '@services/utils/static.data';
+import { rangeCalification, imgsProduct, caracteristicasProductoTabla } from '@services/utils/static.data';
 // css
 import '../product-description/Product.scss';
 
 const Product = () => {
+  // este id servira para buscar el producto en el back
   const { idProduct } = useParams();
 
-  const imgPrueba = {
-    img1: 'https://edgehelmets.com/cdn/shop/files/CAPITAN_AMERICA_ZOOM.jpg?v=1687202275',
-    img2: 'https://encrypted-tbn0.gstatic.com/images?q=tbn:ANd9GcQ8vW2ZGX96EGYANmWC0lf5hMP15jAGuV-K9fb0GQoUhnwLNOybsrAuJBRp_YpPQNzP6Pk&usqp=CAU',
-    img3: 'https://encrypted-tbn0.gstatic.com/images?q=tbn:ANd9GcRO7xhvZAEtfZBmxQGVGHWGJRWBp3I0sFV3SYPn9zqjQofUerr6UYaFBQsyQ_N_lqWt3Z0&usqp=CAU',
-    img4: 'https://encrypted-tbn0.gstatic.com/images?q=tbn:ANd9GcTXSyB0T6VoY4C2hudY6QUdpQTJVt2ppuCyjomDNqUJf20x3SrvQMxzU_FG8jQKdMFGr3E&usqp=CAU'
-  };
-
-  const caracteristicasTabla = [
-    { Marca: 'AGV' },
-    { Línea: 'K1' },
-    { Modelo: 'Pista GP R' },
-    { Norma: 'ECE 22 05' },
-    { Tipo: 'Integral' },
-    { Edad: 'Adulto' }
-  ];
-
-  const [ImgProduct, setImgProduct] = useState(imgPrueba.img1);
-  const [qty, Setqty] = useState(1);
-
-  const showImgProduct = (img) => {
-    setImgProduct(img);
-  };
-
-  const addProduct = () => {
-    Setqty(qty + 1);
-    // aqui hay que agregar logica para que tampoco supere el limite que hay en el stock de la db
-  };
-
-  const minusProduct = () => {
-    if (qty > 1) {
-      Setqty(qty - 1);
-    } else {
-      Setqty(1);
-    }
-  };
+  // state for qty product selected
+  const [qtySelected, SetqtySelected] = useState(1);
 
   return (
-    <div className="bg-container-grid-prueba">
+    <div className="bg-container-product">
       <Container fixed>
-        <div className="container-prueba ">
-          <div className="img-product">
-            <AspectRatio objectFit="contain" minHeight={500}>
-              <img srcSet={ImgProduct} alt="image of product" />
-              <Stack
-                direction={'row'}
-                justifyContent="center"
-                alignItems="center"
-                gap={2}
-                sx={{ marginTop: '.5rem', position: 'absolute', bottom: '5px', left: '50%', right: '50%' }}
-              >
-                <Avatar
-                  // OJO agregar el nombre del producto al ALT
-                  alt="Image product Sharp"
-                  variant="outlined"
-                  src={imgPrueba.img1}
-                  size="lg"
-                  sx={{
-                    cursor: 'pointer',
-                    border: `${ImgProduct === imgPrueba.img1 ? '1px solid var(--red)' : '1px solid #bdbdbd'}`,
-                    '&:hover': { border: '1px solid var(--red)' }
-                  }}
-                  onClick={() => showImgProduct(imgPrueba.img1)}
-                />
-                <Avatar
-                  alt="Remy Sharp"
-                  variant="outlined"
-                  src={imgPrueba.img2}
-                  size="lg"
-                  sx={{
-                    cursor: 'pointer',
-                    border: `${ImgProduct === imgPrueba.img2 ? '1px solid var(--red)' : '1px solid #bdbdbd'}`,
-                    '&:hover': { border: '1px solid var(--red)' }
-                  }}
-                  onClick={() => showImgProduct(imgPrueba.img2)}
-                />
-                <Avatar
-                  alt="Remy Sharp"
-                  variant="outlined"
-                  src={imgPrueba.img3}
-                  size="lg"
-                  sx={{
-                    cursor: 'pointer',
-                    border: `${ImgProduct === imgPrueba.img3 ? '1px solid var(--red)' : '1px solid #bdbdbd'}`,
-                    '&:hover': { border: '1px solid var(--red)' }
-                  }}
-                  onClick={() => showImgProduct(imgPrueba.img3)}
-                />
-                <Avatar
-                  alt="Remy Sharp"
-                  variant="outlined"
-                  src={imgPrueba.img4}
-                  size="lg"
-                  sx={{
-                    cursor: 'pointer',
-                    border: `${ImgProduct === imgPrueba.img4 ? '1px solid var(--red)' : '1px solid #bdbdbd'}`,
-                    '&:hover': { border: '1px solid var(--red)' }
-                  }}
-                  onClick={() => showImgProduct(imgPrueba.img4)}
-                />
-              </Stack>
-            </AspectRatio>
+        <div className="container-grid-product">
+          <div className="img-product-section">
+            <ImageGallery images={imgsProduct} />
+            {/* images se debe pasar las imagenes del product seleccionado  */}
           </div>
-          <aside className="menu">
-            <Card variant="outline" color="neutral" sx={{ height: 'auto', width: '100%' }}>
-              <Stack
-                direction={'row'}
-                justifyContent="space-between"
-                alignItems="center"
-                spacing={2}
-                sx={{
-                  marginBottom: '1rem'
-                }}
-              >
-                <Typography variant="h6">
-                  {idProduct} Lorem ipsum dolor sit amet consectetur adipisicing elit.
-                </Typography>
-                <IconButton
-                  aria-label="Like minimal photography"
-                  size="md"
-                  variant="solid"
-                  color="neutral"
-                  sx={{
-                    borderRadius: '50%',
-                    transform: 'translateY(0%)',
-                    '&:hover': { color: 'red' }
-                  }}
-                >
-                  <Favorite />
-                </IconButton>
-              </Stack>
-
-              <CardContent>
-                <Typography variant="h4" sx={{ marginBottom: '.5rem', fontWeight: 'lighter' }}>
-                  US$ 11.11
-                </Typography>
-
-                <Stack direction="row" justifyContent="flex-start" alignItems="center" spacing={1}>
-                  <Typography variant="caption" sx={{ color: 'grey' }}>
-                    3.0
-                  </Typography>
-                  <Rating name="rating-product" value={3} readOnly />
-                  <Typography variant="caption" sx={{ color: 'grey' }}>
-                    (10)
-                  </Typography>
-                </Stack>
-
-                <Typography variant="body1" sx={{ marginTop: '.5rem' }}>
-                  Envio Gratis
-                </Typography>
-              </CardContent>
-
-              <CardActions orientation="vertical">
-                <Stack
-                  direction="row"
-                  justifyContent="flex-start"
-                  alignItems="center"
-                  spacing={1}
-                  sx={{ marginBottom: '.5rem' }}
-                >
-                  <IconButton size="small" color="primary" onClick={addProduct}>
-                    <AddCircleOutlineIcon />
-                  </IconButton>
-
-                  <Typography variant="body2">{qty}</Typography>
-                  {/* // OJO el max debe ir con respecto a lo que se tenga en la base de datos */}
-
-                  <IconButton size="small" color="primary" disabled={qty === 1 && true} onClick={minusProduct}>
-                    <RemoveCircleOutlineIcon />
-                  </IconButton>
-
-                  <Typography variant="caption" sx={{ color: 'grey' }}>
-                    (10 disponibles)
-                  </Typography>
-                </Stack>
-
-                <Button variant="contained" size="large" sx={{ mb: 2 }} endIcon={<ShoppingCartIcon />}>
-                  Agregar al carrito
-                </Button>
-              </CardActions>
-            </Card>
+          <aside className="card-product-description">
+            <CardProduct
+              titleProduct={'Casco Integral Hro518 Cremer Visor Doble Ahumado'}
+              productPrice={11.11} // debe venir del back
+              avgRate={3} // debe venir del back
+              qtyProduct={10} // debe venir del back
+              qtySelected={qtySelected}
+              SetqtySelected={SetqtySelected}
+            />
           </aside>
-          <div className="menu-card">
-            <Card variant="outline" color="neutral" sx={{ minHeight: 'auto', width: '100%' }}>
-              <Stack
-                direction={'row'}
-                justifyContent="space-between"
-                alignItems="center"
-                spacing={2}
-                sx={{
-                  marginBottom: '1rem'
-                }}
-              >
-                <Typography variant="h6">
-                  {idProduct} Lorem ipsum dolor sit amet consectetur adipisicing elit.
-                </Typography>
-              </Stack>
-
-              <AspectRatio objectFit="contain" minHeight={380}>
-                <img srcSet={ImgProduct} alt="image of product" />
-                <Stack
-                  direction={'row'}
-                  justifyContent="center"
-                  alignItems="center"
-                  gap={2}
-                  sx={{ marginTop: '.5rem', position: 'absolute', bottom: '5px', left: '50%', right: '50%' }}
-                >
-                  <Avatar
-                    // OJO agregar el nombre del producto al ALT
-                    alt="Image product Sharp"
-                    variant="outlined"
-                    src={imgPrueba.img1}
-                    size="lg"
-                    sx={{
-                      cursor: 'pointer',
-                      border: `${ImgProduct === imgPrueba.img1 ? '1px solid var(--red)' : '1px solid #bdbdbd'}`,
-                      '&:hover': { border: '1px solid var(--red)' }
-                    }}
-                    onClick={() => showImgProduct(imgPrueba.img1)}
-                  />
-                  <Avatar
-                    alt="Remy Sharp"
-                    variant="outlined"
-                    src={imgPrueba.img2}
-                    size="lg"
-                    sx={{
-                      cursor: 'pointer',
-                      border: `${ImgProduct === imgPrueba.img2 ? '1px solid var(--red)' : '1px solid #bdbdbd'}`,
-                      '&:hover': { border: '1px solid var(--red)' }
-                    }}
-                    onClick={() => showImgProduct(imgPrueba.img2)}
-                  />
-                  <Avatar
-                    alt="Remy Sharp"
-                    variant="outlined"
-                    src={imgPrueba.img3}
-                    size="lg"
-                    sx={{
-                      cursor: 'pointer',
-                      border: `${ImgProduct === imgPrueba.img3 ? '1px solid var(--red)' : '1px solid #bdbdbd'}`,
-                      '&:hover': { border: '1px solid var(--red)' }
-                    }}
-                    onClick={() => showImgProduct(imgPrueba.img3)}
-                  />
-                  <Avatar
-                    alt="Remy Sharp"
-                    variant="outlined"
-                    src={imgPrueba.img4}
-                    size="lg"
-                    sx={{
-                      cursor: 'pointer',
-                      border: `${ImgProduct === imgPrueba.img4 ? '1px solid var(--red)' : '1px solid #bdbdbd'}`,
-                      '&:hover': { border: '1px solid var(--red)' }
-                    }}
-                    onClick={() => showImgProduct(imgPrueba.img4)}
-                  />
-                </Stack>
-              </AspectRatio>
-
-              <CardContent>
-                <Stack direction="row" justifyContent="space-between" alignItems="center">
-                  <Typography variant="h4" sx={{ marginBottom: '.5rem', fontWeight: 'lighter' }}>
-                    US$ 11.11
-                  </Typography>
-
-                  <IconButton
-                    aria-label="Like minimal photography"
-                    size="md"
-                    variant="solid"
-                    color="neutral"
-                    sx={{
-                      borderRadius: '50%',
-                      transform: 'translateY(0%)',
-                      '&:hover': { color: 'red' }
-                    }}
-                  >
-                    <Favorite />
-                  </IconButton>
-                </Stack>
-
-                <Stack direction="row" justifyContent="flex-start" alignItems="center" spacing={1}>
-                  <Typography variant="caption" sx={{ color: 'grey' }}>
-                    3.0
-                  </Typography>
-                  <Rating name="rating-product" value={3} readOnly />
-                  <Typography variant="caption" sx={{ color: 'grey' }}>
-                    (10)
-                  </Typography>
-                </Stack>
-
-                <Typography variant="body1" sx={{ marginTop: '.5rem' }}>
-                  Envio Gratis
-                </Typography>
-              </CardContent>
-
-              <CardActions orientation="vertical" className="action-card" sx={{ padding: '18px', zIndex: 10 }}>
-                <Stack direction="row" justifyContent="space-evenly" alignItems="center" spacing={1}>
-                  <Stack direction="column">
-                    <Stack
-                      direction="row"
-                      justifyContent="flex-start"
-                      alignItems="center"
-                      spacing={1}
-                      sx={{ marginBottom: '.5rem' }}
-                    >
-                      <IconButton size="small" color="primary" onClick={addProduct}>
-                        <AddCircleOutlineIcon />
-                      </IconButton>
-
-                      <Typography variant="body2">{qty}</Typography>
-                      {/* // OJO el max debe ir con respecto a lo que se tenga en la base de datos */}
-
-                      <IconButton size="small" color="primary" disabled={qty === 1 && true} onClick={minusProduct}>
-                        <RemoveCircleOutlineIcon />
-                      </IconButton>
-                    </Stack>
-                    <Typography variant="caption" sx={{ color: 'grey' }}>
-                      (10 disponibles)
-                    </Typography>
-                  </Stack>
-
-                  <Button variant="contained" size="large" sx={{ mb: 2 }} endIcon={<ShoppingCartIcon />}>
-                    Agregar al carrito
-                  </Button>
-                </Stack>
-              </CardActions>
-            </Card>
+          <div className="card-product-description-adaptive">
+            <CardProductAdaptive
+              titleProduct={'Casco Integral Hro518 Cremer Visor Doble Ahumado'}
+              productPrice={11.11} // debe venir del back
+              avgRate={3} // debe venir del back
+              qtyProduct={10} // debe venir del back
+              imgsProduct={imgsProduct} // estass deben ser las imgs del product de la db
+              qtySelected={qtySelected}
+              SetqtySelected={SetqtySelected}
+            />
           </div>
-          <div className="info">
-            <Accordion defaultExpanded={true}>
-              <AccordionSummary expandIcon={<ExpandMoreIcon />} sx={{ fontSize: '20px' }}>
-                Informacion del producto
-              </AccordionSummary>
-              <AccordionDetails sx={{ width: '100%', margin: 'auto' }}>
-                <Typography variant="subtitle1" sx={{ marginBottom: '.5rem' }}>
-                  Caracteristicas
-                </Typography>
-
-                <Box
-                  sx={{
-                    display: 'flex',
-                    flexDirection: 'column',
-                    justifyContent: 'center',
-                    width: '60%',
-                    height: 'auto',
-                    margin: 'auto'
-                  }}
-                >
-                  <List>
-                    {caracteristicasTabla.map((item, index) => {
-                      const [feature, value] = Object.entries(item)[0];
-                      return (
-                        <ListItem
-                          key={index}
-                          sx={{
-                            display: 'flex',
-                            justifyContent: 'space-between',
-                            alignItems: 'center',
-                            background: `${index % 2 === 0 ? '#E6E6E6' : 'none'}`,
-                            overflow: 'hidden'
-                          }}
-                        >
-                          <ListItemText sx={{ width: '50%', height: 'auto' }}>
-                            <Typography variant="body2">{feature}</Typography>
-                          </ListItemText>
-                          <ListItemText sx={{ width: '50%', height: 'auto' }}>
-                            <Typography variant="body2" sx={{ fontWeight: 'lighter' }}>
-                              {value}
-                            </Typography>
-                          </ListItemText>
-                        </ListItem>
-                      );
-                    })}
-                  </List>
-                </Box>
-
-                <Divider sx={{ margin: '2rem 0rem' }} />
-
-                <Typography variant="body1" sx={{ marginBottom: '.5rem' }}>
-                  Descripcion
-                </Typography>
-
-                <Typography variant="body2" align="justify" sx={{ fontWeight: 'lighter' }}>
-                  Contrary to popular belief, Lorem Ipsum is not simply random text. It has roots in a piece of
-                  classical Latin literature from 45 BC, making it over 2000 years old. Richard McClintock, a Latin
-                  professor at Hampden-Sydney College in Virginia, looked up one of the more obscure Latin words,
-                  consectetur, from a Lorem Ipsum passage, and going through the cites of the word in classical
-                  literature, discovered the undoubtable source. Lorem Ipsum comes from sections 1.10.32 and 1.10.33 of
-                  de Finibus Bonorum et Malorum (The Extremes of Good and Evil) by Cicero, written in 45 BC. This book
-                  is a treatise on the theory of ethics, very popular during the Renaissance. The first line of Lorem
-                  Ipsum, Lorem ipsum dolor sit amet.., comes from a line in section 1.10.32. The standard chunk of Lorem
-                  Ipsum used since the 1500s is reproduced below for those interested. Sections 1.10.32 and 1.10.33 from
-                  de Finibus Bonorum et Malorum by Cicero are also reproduced in their exact original form, accompanied
-                  by English versions from the 1914 translation by H. Rackham.
-                </Typography>
-              </AccordionDetails>
-            </Accordion>
+          <div className="info-product-section">
+            <AccordionInfoProduct
+              characteristicsProduct={caracteristicasProductoTabla} // esta deben venir de la db del product
+              descriptionProduct={
+                'Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industrys standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it to make a type specimen book. It has survived not only five centuries, but also the leap into electronic typesetting, remaining essentially unchanged. It was popularised in the 1960s with the release of Letraset sheets containing Lorem Ipsum passages, and more recently with desktop publishing software like Aldus PageMaker including versions of Lorem Ipsum.'
+              }
+            />
           </div>
-          <div className="pregunta">
-            <Accordion defaultExpanded={true}>
-              <AccordionSummary expandIcon={<ExpandMoreIcon />} sx={{ fontSize: '20px' }}>
-                Preguntas y respuestas
-              </AccordionSummary>
-              <AccordionDetails sx={{ width: '100%', margin: 'auto' }}>
-                <Box
-                  component="form"
-                  direction={'row'}
-                  sx={{ marginY: 2, width: '100%', display: 'flex', justifyContent: 'space-evenly' }}
-                >
-                  <TextField
-                    id="input-question"
-                    value=""
-                    variant="outlined"
-                    type="text"
-                    placeholder="Haz una pregunta"
-                    sx={{ width: '70%' }}
-                  />
-
-                  <Button type="submit" variant="outlined" size="md" width={'30%'}>
-                    Preguntar
-                  </Button>
-                </Box>
-
-                <Box
-                  sx={{
-                    display: 'flex',
-                    flexDirection: 'column',
-                    justifyContent: 'center',
-                    width: '100%',
-                    height: 'auto',
-                    margin: 'auto'
-                  }}
-                >
-                  <List>
-                    {/* pregunta */}
-                    <ListItem>
-                      <ListItemText>
-                        <Typography variant="body2">
-                          Lorem ipsum dolor sit amet consectetur, adipisicing elit?
-                        </Typography>
-                      </ListItemText>
-                    </ListItem>
-                    {/* respuesta */}
-                    <ListItem>
-                      <ListItemIcon>
-                        <SubdirectoryArrowRightIcon />
-                      </ListItemIcon>
-                      <ListItemText>
-                        <Typography variant="body2" sx={{ fontWeight: 'lighter' }}>
-                          Lorem ipsum, dolor sit amet consectetur adipisicing elit. Eligendi quaerat laboriosam
-                          similique modi voluptates voluptate ut aperiam provident
-                        </Typography>
-                      </ListItemText>
-                    </ListItem>
-                  </List>
-                </Box>
-              </AccordionDetails>
-            </Accordion>
+          <div className="question-and-answers-section">
+            <AccordionQA />
           </div>
-          <div className="calificacion">
-            <Accordion defaultExpanded={true}>
-              <AccordionSummary expandIcon={<ExpandMoreIcon />} sx={{ fontSize: '20px' }}>
-                Calificaciones
-              </AccordionSummary>
-              <AccordionDetails sx={{ width: '250px', height: 'auto', margin: 'auto' }}>
-                <List>
-                  {rangeCalification.map((item) => (
-                    <ListItem key={item.id} alignItems="center">
-                      <ListItemIcon>
-                        <Rating name="rating-product" value={item.value} readOnly />
-                      </ListItemIcon>
-
-                      <ListItemText>
-                        <Typography sx={{ fontSize: '14px', fontWeight: 400, color: 'grey', marginLeft: 1 }}>
-                          {' '}
-                          ({item.mount})
-                        </Typography>
-                      </ListItemText>
-                    </ListItem>
-                  ))}
-                </List>
-              </AccordionDetails>
-            </Accordion>
+          <div className="calification-product-section">
+            <AccordionRateProduct rangeCalification={rangeCalification} />
           </div>
         </div>
