+++ tony-shop/src/molecules/card/cardProduct-details/CardProduct-adaptive/CardProductAdaptive.scss
@@ -1,0 +1,7 @@
+.card-action-section {
+  position: fixed;
+  bottom: 0;
+  left: 0;
+  right: 0;
+  background-color: #fff;
+}
